Add files via upload
Created some framework for the database. Created an events table and an accounts table. This is the dump of the database. In order to use the database run the following commands in the terminal
mysqladmin create simple_planner
mysql simple_planner < simple_planner.sql

When you are done modifying the database use the following command (might need root privileges)
mysqldump --databases simple_planner -p > simple_planner.sql
and then commit the changed file